wallet work completed

+++ sonichub/order_managements/views.py
@@ -1,8 +1,8 @@
 from django.shortcuts import render, redirect, get_object_or_404
-from user_panel.models import Address
+from user_panel.models import Address,Transaction
 from cart_management.models import Cart
 from order_managements.models import Order_Main_data, Order_Sub_data
-from user_panel.models import Address
+from user_panel.models import Address,Transaction
 from user_authentication.models import UserProfile
 from django.utils import timezone
 from product_management.models import Product_Variant
@@ -15,14 +15,22 @@ import razorpay
 from django.conf import settings
 from coupon_management.models import Coupon, Users_Coupon
 from django.contrib import messages
+from django.db.models import Sum
 
 
 
+# payment success function for online order
 def payment_success(request, order_id):
     order_id = "#" + order_id
     main_order_obj = Order_Main_data.objects.get(order_id=order_id)
     main_order_obj.payment_status = True
     main_order_obj.save()
+
+    id = request.user.id
+    #clearing the cart after payment
+    data = Cart.objects.filter(user=id)
+    data.delete()
+
     return render(request, "user_side/order-success.html", {"order_id": order_id})
 
 
@@ -39,12 +47,28 @@ def online_payment(request, order_id):
     return render(request, "user_side/online-payment-summary.html", context)
 
 
+
 def cancel_order(request, order_id, user_id):
     if request.method == "POST":
         data = Order_Main_data.objects.get(order_id=order_id)
         data.order_status = "Cancelled"
         data.save()
+
+        if data.payment_option == 'online payment' or data.payment_option == 'wallet payment':
+            amount = request.POST.get('total_amount')
+            user = UserProfile.objects.get(id=user_id)
+            
+            transaction_data = Transaction.objects.create(
+                user=user,
+                description="Cancelled Order " +" "+ order_id,
+                amount=amount,
+                transaction_type="Credit",
+               
+            )
+
+        print(user_id)
         return redirect("order:order-list", user_id)
+
 
 
 @cache_control(no_cache=True, must_revalidate=True, no_store=True)
@@ -60,7 +84,9 @@ def order_details(request, id):
 def order_list(request, id):
     print(id)
 
-    order_main_data = Order_Main_data.objects.filter(user_id=id).order_by('id').reverse()
+    order_main_data = (
+        Order_Main_data.objects.filter(user_id=id).order_by("id").reverse()
+    )
 
     grouped_data = defaultdict(list)
 
@@ -89,36 +115,62 @@ def confirm_order(request, id):
         data = Cart.objects.filter(user=id)
         if not data:
             return redirect("cart:product-cart")
-        
+
+        # finding total
         sub_total = 0
 
         for i in data:
             if i.variant.variant_status:
                 sub_total += i.quantity * i.product.offer_price
 
-
         order_status = request.POST.get("order_status")
         address = request.POST.get("addressId")
         payment_option = request.POST.get("payment_option")
         total_price = request.POST.get("total_price")
-        coupon_id = request.POST.get("coupon_id")
-        
+        coupon_code = request.POST.get("coupon_code")
+
+
+
+        # coupon recheck...
+        if coupon_code is not None and coupon_code != "None":
+            try:
+                coupon = Coupon.objects.get(
+                    Coupon_code=coupon_code,
+                    is_active=True,
+                    expiry_date__gte=timezone.now().date(),
+                    minimum_amount__lt=sub_total,
+                )
+
+                if sub_total > coupon.minimum_amount:
+                    discount = float(coupon.discount)
+                    amount = float(total_price) * (discount / 100)
+                    total_price = round(float(total_price) - amount, 2)
+
+                else:
+                    messages.warning(request, "Invalid coupon")
+                    return redirect("order:checkout", user)
+
+            except Exception as e:
+                print(e)
+                messages.warning(request, "Enter valid coupon")
+                return redirect("order:checkout", user)
 
         current_date = timezone.now().date()
         formatted_date = current_date.strftime("%Y%m%d")
         r_string = get_random_string(length=2)
         address_id = Address.objects.get(id=address)
         user_id = UserProfile.objects.get(id=id)
-        order_id = "#" + str(formatted_date) + str(user) + r_string.upper()
-       
+
         main_order = Order_Main_data.objects.create(
             total_amount=total_price,
             order_status=order_status,
             address=address_id,
             user=user_id,
-            payment_option=payment_option,
-            order_id=order_id,
+            payment_option=payment_option
         )
+        order_id = "#" + str(formatted_date) + str(user) + str(main_order.id)
+        main_order.order_id = order_id
+        main_order.save()
 
         main_order_id = Order_Main_data.objects.get(id=main_order.id)
 
@@ -128,18 +180,30 @@ def confirm_order(request, id):
                 main_order=main_order_id,
                 user=user_id,
                 variant=product.variant,
+
             )
+
+   
+        if payment_option == "online payment":
+            return redirect("order:online-payment", order_id)
+        
+        if payment_option == 'wallet payment':
+            return redirect("user_panel:wallet-payment",order_id, id)
+        
+
+        # clearing the cart
         data.delete()
 
-        total_price_float = float(total_price)
-
-        
         # changing coupon status
-        if coupon_id is not None and coupon_id != "None":
-            coupon = Coupon.objects.get(id=coupon_id)
+        if coupon_code is not None and coupon_code != "None":
+            coupon = Coupon.objects.get(Coupon_code=coupon_code)
             coupon.is_active = False
             coupon.save()
 
+
+        # creating razorpay object
+            
+        total_price_float = float(total_price)
 
         client = razorpay.Client(
             auth=(settings.RAZORPAY_KEY, settings.RAZORPAY_KEY_SECRET)
@@ -154,17 +218,14 @@ def confirm_order(request, id):
         main_order_id.payment_id = payment["id"]
         main_order_id.save()
 
-        if payment_option == "online payment":
-            return redirect("order:online-payment", order_id)
+
         
         main_order_id.payment_status = True
         main_order_id.save()
-        
 
-        context = {"order_id": order_id, "address": address_id,"subtotal":sub_total}
+        context = {"order_id": order_id, "address": address_id, "subtotal": sub_total}
 
         return render(request, "user_side/order-success.html", context)
-
 
 
 
@@ -172,8 +233,8 @@ def confirm_order(request, id):
 def checkout(request, id):
     data = Cart.objects.filter(user=id)
     total_price = 0
-    sub_total =0
-    coupon_id =None
+    sub_total = 0
+    coupon_code = None
     for i in data:
         if i.variant.variant_status:
             total_price += i.quantity * i.product.offer_price
@@ -188,15 +249,15 @@ def checkout(request, id):
                 is_active=True,
                 expiry_date__gte=timezone.now().date(),
             )
-            
+
             if total_price > coupon.minimum_amount:
                 discount = float(coupon.discount)
                 amount = float(total_price) * (discount / 100)
-                total_price = round(float(total_price) - amount, 2) 
+                total_price = round(float(total_price) - amount, 2)
                 coupon_id = coupon.id
 
             else:
-                messages.warning(request,"Enter valid coupon")
+                messages.warning(request, "Invalid coupon")
 
         except Exception as e:
             print(e)
@@ -206,8 +267,7 @@ def checkout(request, id):
         "address": Address.objects.filter(user=id, status=True),
         "products": data,
         "total_price": total_price,
-        "coupon_id":coupon_id,
-        "subtotal":sub_total
-        
+        "coupon_code": coupon_code,
+        "subtotal": sub_total,
     }
     return render(request, "user_side/checkout.html", content)
